- 路线 A：四个 Session（AR/EOS/SAR/SBIRS）可选注册 LifecycleRecorder，
  StepWithResult 内部自动驱动，消除调用方漏调失步陷阱
- 三写约束统一（abort_reason + diagnostics + 日志），SBIRS powered-off 收尾
- DebugView → JSON 序列化参考实现（examples 各模块目录，规则 12）
- 三写守卫测试（5 模块）+ EOS powered-off status 修复
- replay codec decode 白名单 + AR abort status 显式化
- 修复 sar_session_usage 坏示例（真实链路预算 SNR 门限）

+++ tests/contract/public_api/public_headers_smoke_test.cpp
@@ -455,7 +455,7 @@ TEST(PublicHeadersSmokeTest, SarPublicSurfaceSupportsMinimalUsage) {
   EXPECT_TRUE(session.TryApplyRuntimeConfig(patch));
 
   const session::SarCycleResult result = session.StepWithResult(input);
-  EXPECT_TRUE(result.executed_this_cycle) << result.abort_reason;
+  EXPECT_TRUE(result.executed_this_cycle) << static_cast<int>(result.abort_reason);
   EXPECT_EQ(result.output_frame.range_sample_count, 64U);
   EXPECT_TRUE(result.output_frame.has_raw_echo);
   EXPECT_TRUE(result.output_frame.has_range_compressed_echo);
